requirements.txt
python-dotenv==1.0.1
Requests==2.32.3
json==2.0.9
pandas==2.1.2
numpy==1.26.2